+++ CHANGELOG.md
@@ -21,6 +21,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   - Plugin Messages / Translations
 - `region` and `wand` conditions now supports variables
 - `JobsReborn`, `Quests`, `McMMO` now supports variables
+- notification categories `conversation_start` and `conversation_end`
 ### Changed
 - Spigot is no longer supported, paper is now required 
 - message.yml file was deleted and instead the lang folder now contains all translations
@@ -32,10 +33,12 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - `%quester%` variable is used instead of `%npc%` for the quester's name in conversations; `%npc` variable is used now for the new Npc system
 - `fish` objective use block selector instead of a QuestItem
 - everything that used `Citizens` NPC ids now uses the BQ NpcIDs
+- conversations now always print their start and stop conversation message
 ### Deprecated
 ### Removed
 - undocumented prefix feature in conversation
 - `citizens_npcs_by_name` config option, which is now part of the id
+- start and stop sound in the configuration for conversations were removed in favor of the notification system that now also has the two new build in categories `conversation_start` and `conversation_end`
 ### Fixed
 - Reloading plugin did not reload Menu config
 - potions generated not extended/upgraded since 2.2.1